Update Script
FIXED: When there was no email that met the deletion criteria, the script would create an exception, and an email would be sent as a summary of failures for Google Apps Scripts. Now, it will simply count the number of emails to be deleted, and if it equals 0, the script will end.